refactor proj
