+++ app/src/main/java/com/example/consecutivepractice/screens/GameFilterScreen.kt
@@ -1,4 +1,4 @@
-package com.example.consecutivepractice
+package com.example.consecutivepractice.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -32,6 +32,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.example.consecutivepractice.viewmodels.GameFilterViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
