+++ app/src/main/java/com/example/consecutivepractice/screens/PlaceholderScreen.kt
@@ -1,4 +1,4 @@
-package com.example.consecutivepractice
+package com.example.consecutivepractice.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
