--- app/src/main/java/com/example/consecutivepractice/ui/Computer.kt
@@ -1,66 +1,0 @@
-package com.example.consecutivepractice.ui
-
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.PathFillType
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.StrokeJoin
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.graphics.vector.path
-import androidx.compose.ui.unit.dp
-
-val Computer: ImageVector
-    get() {
-        if (_Computer != null) {
-            return _Computer!!
-        }
-        _Computer = ImageVector.Builder(
-            name = "Computer",
-            defaultWidth = 24.dp,
-            defaultHeight = 24.dp,
-            viewportWidth = 960f,
-            viewportHeight = 960f
-        ).apply {
-            path(
-                fill = SolidColor(Color.Black),
-                fillAlpha = 1.0f,
-                stroke = null,
-                strokeAlpha = 1.0f,
-                strokeLineWidth = 1.0f,
-                strokeLineCap = StrokeCap.Butt,
-                strokeLineJoin = StrokeJoin.Miter,
-                strokeLineMiter = 1.0f,
-                pathFillType = PathFillType.NonZero
-            ) {
-                moveTo(40f, 840f)
-                verticalLineToRelative(-80f)
-                horizontalLineToRelative(880f)
-                verticalLineToRelative(80f)
-                close()
-                moveToRelative(120f, -120f)
-                quadToRelative(-33f, 0f, -56.5f, -23.5f)
-                reflectiveQuadTo(80f, 640f)
-                verticalLineToRelative(-440f)
-                quadToRelative(0f, -33f, 23.5f, -56.5f)
-                reflectiveQuadTo(160f, 120f)
-                horizontalLineToRelative(640f)
-                quadToRelative(33f, 0f, 56.5f, 23.5f)
-                reflectiveQuadTo(880f, 200f)
-                verticalLineToRelative(440f)
-                quadToRelative(0f, 33f, -23.5f, 56.5f)
-                reflectiveQuadTo(800f, 720f)
-                close()
-                moveToRelative(0f, -80f)
-                horizontalLineToRelative(640f)
-                verticalLineToRelative(-440f)
-                horizontalLineTo(160f)
-                close()
-                moveToRelative(0f, 0f)
-                verticalLineToRelative(-440f)
-                close()
-            }
-        }.build()
-        return _Computer!!
-    }
-
-private var _Computer: ImageVector? = null
