+++ app/src/main/java/com/example/consecutivepractice/viewmodels/GameFilterViewModel.kt
@@ -1,8 +1,9 @@
-package com.example.consecutivepractice
+package com.example.consecutivepractice.viewmodels
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.consecutivepractice.repositories.FilterRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
